Compilation fixes

+++ sw/airborne/subsystems/datalink/datalink.h
@@ -67,7 +67,7 @@ EXTERN bool datalink_enabled;
 
 /** Convenience macro to fill dl_buffer */
 #define DatalinkFillDlBuffer(_buf, _len) { \
-  uint8_t _i = 0; \
+  uint16_t _i = 0; \
   for (_i = 0; _i < _len; _i++) { \
     dl_buffer[_i] = _buf[_i]; \
   } \
